Cambio models ranking

+++ appresultados/models.py
@@ -7,7 +7,7 @@ from appweb.models import Usuario
 class Resultado(models.Model):
     cuestionario = models.ForeignKey(Cuestionario, on_delete=models.CASCADE)
     usuario = models.ForeignKey(Usuario, on_delete=models.CASCADE)
-    puntaje = models.FloatField()
+    puntaje = models.DecimalField(max_digits=10,decimal_places=2)
 
     def __str__(self):
         return str(self.pk)
